requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2
chromadb>=0.4.18
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
textblob>=0.17.0
vaderSentiment>=3.3.0
plotly>=5.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0 